Re-merge DMF-LUA and DMF with ifdefs
